Textures

+++ VoxelGame.Core/Data/Graphics/IIndexBuffer.cs
@@ -3,6 +3,6 @@ namespace VoxelGame.Core.Data.Graphics;
 public interface IIndexBuffer
 {
     
-    public void Fetch(ushort[] indices);
+    public void Fetch(uint[] indices);
     
 }
